dos2unix

+++ info_linux.go
@@ -17,7 +17,7 @@ func GetFileInfo(path string, info fs.FileInfo) *FileInfo {
 		Path:         path,
 		Size:         info.Size(),
 		IsDir:        info.IsDir(),
-		CreatedTime:  time.Unix(stat.Ctim.Sec, stat.Ctim.Nsec),
+		CreatedTime:  time.Unix(int64(stat.Ctim.Sec), int64(stat.Ctim.Nsec)),
 		ModifiedTime: info.ModTime(),
 	}
 }
